Redesign!
Found a cool looking free theme on HTML5 UP. Pulled it in and tried to
figure out how to get it all working in Rails. The theme uses the skel
javascript/css grid library, which I'm honestly not a huge fan of. It's
kind of a pain to get in there, but I think I did an ok job. Not sure we
really need all the different responsive styles, might want to remove
some of them at some point. Also killed a bunch of IE 8 support in the
theme because it was being a pain.