Add Evil Lairs section
Not to be limited to only people, and since we want to highlight actual
evil things, I added an Evil Lairs section to house links to
particularly evil code. Seeded with two repos -- one that I found as a
suggestion on twitter, the other as my favorite troll repo. This means
we can solicit many more evil lairs for addition from the general
populace.